+++ Modules/Administration/Database/Migrations/2023_10_13_192756_create_product_types_table.php
@@ -21,14 +21,18 @@ return new class extends Migration
             $table->unsignedBigInteger('users_id');
             $table->boolean('is_synced')->default(0);
             $table->timestamp('deleted_at')->nullable();
-
-            $table->foreign('users_id')
-                ->references('id')
-                ->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
             $table->timestamps();
         });
+
+        if (Schema::hasTable('product_types') && Schema::hasTable('users')) {
+            Schema::table('product_types', function (Blueprint $table) {
+                $table->foreign('users_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
